+++ Sources/StreamChat/Query/Filter.swift
@@ -6,7 +6,7 @@ import Foundation
 
 /// An enum with possible operators to use in filters.
 public enum FilterOperator: String {
-    /// Matches values that are equal to a specified value.
+    /// Matches values that are equal to a specified value or matches all of the values in an array.
     case equal = "$eq"
 
     /// Matches all values that are not equal to a specified value.
@@ -294,6 +294,17 @@ public extension Filter {
             operator: .equal,
             key: key,
             value: value,
+            valueMapper: key.valueMapper,
+            keyPathString: key.keyPathString
+        )
+    }
+    
+    /// Matches values that are equal to a specified values.
+    static func equal<Value: Encodable>(_ key: FilterKey<Scope, Value>, values: [Value]) -> Filter {
+        .init(
+            operator: .equal,
+            key: key,
+            value: values,
             valueMapper: key.valueMapper,
             keyPathString: key.keyPathString
         )